[#25] refactor: WeekDays 파일 분리
