+++ iOS/Airbnb/Airbnb/Accommodations/ConditionSetting/DatePicker/Paging/PagingDataSource.swift
@@ -11,6 +11,7 @@ import UIKit
 final class PagingDataSource: NSObject, UICollectionViewDataSource {
     
     private let numberOfMonths: Int = 6
+    // 현재 월
     
     // MARK: - Methods
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
@@ -21,7 +22,9 @@ final class PagingDataSource: NSObject, UICollectionViewDataSource {
         guard let cell = collectionView
             .dequeueReusableCell(withReuseIdentifier: PagingCell.identifier,
                                  for: indexPath) as? PagingCell else { return UICollectionViewCell() }
-
+        // 현재 월 + 인덱스를 하면
         return cell
     }
 }
+
+
