+++ iOS/Airbnb/Airbnb/Accommodations/ConditionSetting/DatePicker/WeekDayView.swift
@@ -10,41 +10,6 @@ import UIKit
 
 final class WeekDayView: UIStackView {
     
-    enum WeekDays: CustomStringConvertible, CaseIterable {
-        
-        case Sun, Mon, Tue, Wed, Thu, Fri, Sat
-
-        var description: String {
-            switch self {
-            case .Sun:
-                return "Sun"
-            case .Mon:
-                return "Mon"
-            case .Tue:
-                return "Tue"
-            case .Wed:
-                return "Wed"
-            case .Thu:
-                return "Thu"
-            case .Fri:
-                return "Fri"
-            case .Sat:
-                return "Sat"
-            }
-        }
-        
-        func color() -> UIColor {
-            switch self {
-            case .Sun:
-                return .systemRed
-            case .Sat:
-                return .systemBlue
-            default:
-                return .systemGray
-            }
-        }
-    }
-    
     // MARK: - Lifecycle
     override init(frame: CGRect) {
         super.init(frame: frame)
